+++ pysat/constellations/test_diff_same.py
@@ -7,8 +7,8 @@ test instruments
 
 inst1 = pysat.Instrument('pysat', 'testing',
                          clean_level='clean',
-                         sat_id='6000')
+                         inst_id='6000')
 inst2 = pysat.Instrument('pysat', 'testing',
                          clean_level='clean',
-                         sat_id='6000')
+                         inst_id='6000')
 instruments = [inst1, inst2]