+++ pysat/constellations/test_diff_similar.py
@@ -8,8 +8,8 @@ instruments
 inst1 = pysat.Instrument('pysat', 'testing',
                          clean_level='clean',
                          tag='mlt_offset',
-                         sat_id='6000')
+                         inst_id='6000')
 inst2 = pysat.Instrument('pysat', 'testing',
                          clean_level='clean',
-                         sat_id='6000')
+                         inst_id='6000')
 instruments = [inst1, inst2]