+++ CHANGELOG.md
@@ -64,6 +64,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Add tests for acknowledgements and references
   - Removed implicit conversion to integers in
     methods.general.convert_timestamp_to_datetime
+  - Renamed `sat_id` Instrument keyword argument to `inst_id`
 
 ## [2.2.1] - 2020-07-29
 - Documentation